👌 IMPROVE: moved components to components folder
